Add modId option
This option allows to specify a mod ID instead of the
full language directory path for ease of use and setup.

In other words, the value of this new option will be
substituted in a pre-determined file path (see
documentation in case you're wondering which one) and
that path will be used instead.

Of course, the path can always be specified manually

Signed-off by TheSilkMiner