Add Getting Started guide
(documentation-updater-commit)

With the release of the version 1.1, a lot of things have changed. With the
addition of this little "Getting Started" guide, a new user will be able to
start using this plug-in in no time. And a quicker start means one more
customer (I don't remember who said this...)

Signed-off by TheSilkMiner <sir.karati@outlook.com>