Add SKL Interpreter to project
This way, we can start moving all the various bits and pieces to a
secondary file, without having to add various properties to tasks.
Everything will be performed through a single file and a single
parser, with the addition of all the various structures and default
values from a central location.

Signed-off by TheSilkMiner <sir.karati@outlook.com>