Add SKD-based task description and remove old properties
This is the main change of version 2.0

This has mainly been done to allow a more flexible and centralized
way to provide task descriptions. Instead of having various
specifications for every task, now the user can specify everything
in a single file and supply it to the task.

In the future, some more changes will be apported to the format so
that a single file will be able to support multiple task descriptions
(which is the real goal of the file) through IDs, specified directly
in the task initialization itself.

Since not all users may be familiar and/or able to construct a
database descriptors, a friendly GUI creation tool is in the
works: it will be run just by double clicking the plugin. While this
feature is being developed, we decided to help the end user by
providing both a default and an example descriptor files, which can
be found under the script's assets
(src/main/resources/assets/database/** for the non-techie). You are
free to edit the file as much as you like, so don't worry.

To the developers: to develop with this plugin, you can download
every jar you'd like, but I suggest you to use the minimal jar, due
to its more compact size. In case you want to test something though
and to do this you need to run the plugin, the usage of the normal
jar is mandatory because it contains the SKL interpreter used to
read the database and thus run the task.

Also, if you want to contribute, please keep the two codebases
separated: groovy files in the groovy directory and java files in
the java directory. Add methods to the bridge classes to "talk"
between the two code-bases. My OCD will thank you.

Signed-off by TheSilkMiner <sir.karati@outlook.com>